* upstream/master: (47 commits)
  install-manual: fix sec. swarm manager connect
  fix bug that http server should return when handler is nil
  fix set reschedule policy docs
  hide error in info when empty
  Krane project has been closed in favor of Machine
  fix typos in swarm/cluster.go
  Updating Release checklist
  Bump version to 1.2.2-rc1
  Prevent double RLock() of RWMutex. It may lead to deadlock if there is Lock() call from another goroutine in between.
  Bump version to v1.2.1
  Carry over pr/2137 with minor updates.
  Filter for containerslots-label
  use --cpu-shares instead of -c in integration test
  Fix api/ps tests.
  fix typos in test and other files
  use engine-api to replace dockerclient
  Update stats test to stop connection
  Update Edits
  Edit typos in 'Try Swarm at scale'
  convert docs Dockerfiles to use docs/base:oss
  ...

+++ cluster/cluster.go
@@ -79,7 +79,7 @@ type Cluster interface {
 	TotalMemory() int64
 
 	// Return the number of CPUs in the cluster
-	TotalCpus() int
+	TotalCpus() int64
 
 	// Register an event handler for cluster-wide events.
 	RegisterEventHandler(h EventHandler) error
